feat: add login user api with basic-auth
