+++ server/httpserver/root.go
@@ -2,24 +2,23 @@ package main
 
 import (
 	"fmt"
-	"github.com/salehborhani/todo-list/server/httpserver/request"
-	"log"
+	"github.com/salehborhani/todo-list/server/httpserver/handler"
 	"net/http"
 )
 
 func main() {
 
-	// TODO - register user with name, email and password
-	http.HandleFunc("/api/register/", request.RegisterUser)
+	http.HandleFunc("/api/register/", handler.RegisterUser)
+
+	// TODO - login user with the creds
+
+	http.HandleFunc("/api/login/", handler.LoginUser)
+
 	server := &http.Server{
 		Addr: ":8080",
 	}
 	fmt.Println("Starting server on", server.Addr)
-
 	if err := server.ListenAndServe(); err != nil {
-		log.Fatalln(err)
+
 	}
-
-	// TODO - login user with the creds
-
 }
